+++ src/tests/work-item/work-item-list/displayAndUpdateWorkItemDetails.spec.js
@@ -14,7 +14,7 @@
  * @author ldimaggi
  */
 
-var WorkItemListPage = require('./work-item-list.page'),
+var WorkItemListPage = require('./page-objects/work-item-list.page'),
   testSupport = require('./testSupport');
 
 describe('Work item list', function () {
