+++ src/tests/work-item/work-item-list/page-objects/notifications.page.js
@@ -12,16 +12,16 @@
  * Chat Page Definition - placeholder
  */
 
-let testSupport = require('./testSupport');
+let testSupport = require('../testSupport');
 let CommonPage = require('./common.page');
-let constants = require("./constants");
+let constants = require("../constants");
 let until = protractor.ExpectedConditions;
 
-class ChatPage {
+class NotificationsPage {
 
  constructor(login) {
  };
 
 }
 
-module.exports = ChatPage;
+module.exports = NotificationsPage;
