+++ src/tests/work-item/work-item-list/page-objects/settings.page.js
@@ -9,19 +9,19 @@
 'use strict';
 
 /*
- * Chat Page Definition - placeholder
+ * Code Page Definition - placeholder
  */
 
-let testSupport = require('./testSupport');
+let testSupport = require('../testSupport');
 let CommonPage = require('./common.page');
-let constants = require("./constants");
+let constants = require("../constants");
 let until = protractor.ExpectedConditions;
 
-class NotificationsPage {
+class SettingsPage {
 
  constructor(login) {
  };
 
 }
 
-module.exports = NotificationsPage;
+module.exports = SettingsPage;
