+++ src/tests/work-item/work-item-list/page-objects/test.page.js
@@ -9,19 +9,19 @@
 'use strict';
 
 /*
- * Hypothesis Page Definition - placeholder
+ * Test Page Definition - placeholder
  */
 
-let testSupport = require('./testSupport');
+let testSupport = require('../testSupport');
 let CommonPage = require('./common.page');
-let constants = require("./constants");
+let constants = require("../constants");
 let until = protractor.ExpectedConditions;
 
-class HypothesisPage {
+class TestPage {
 
  constructor(login) {
  };
 
 }
 
-module.exports = HypothesisPage;
+module.exports = TestPage;
