+++ src/tests/work-item/work-item-list/page-objects/work-item-board.page.js
@@ -13,9 +13,9 @@
  * Work Item Board Page Definition
  */
 
-let testSupport = require('./testSupport');
+let testSupport = require('../testSupport');
 let CommonPage = require('./common.page');
-let constants = require("./constants");
+let constants = require("../constants");
 let until = protractor.ExpectedConditions;
 
 class WorkItemBoardPage {
